Extracted useConstraint types to a separate file
